requirements.txt
Flask
psycopg2-binary
python-dotenv
werkzeug
youtube-search-python
httpx==0.27.2
flask-cors
